UserController로 기능 병합
