+++ src/main/java/com/example/demo/controller/UserController.java
@@ -5,10 +5,18 @@ import java.util.regex.Pattern;
 
 import com.example.demo.dao.UserDAO;
 import com.example.demo.dto.User;
+import com.example.demo.security.JwtResponse;
+import com.example.demo.security.JwtTokenUtil;
+import com.example.demo.security.JwtUserDetailsService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.DisabledException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,19 +36,24 @@ public class UserController {
     @Autowired
     PasswordEncoder passwordEncoder;
 
-
+    @Autowired
+    private AuthenticationManager authenticationManager;
+    
+	@Autowired
+    private JwtTokenUtil jwtTokenUtil;
+    
+	@Autowired
+    private JwtUserDetailsService userDetailsService;
 
     @PostMapping(value = "/login")
-    public Object Login(@RequestBody User loginUser) {
-        Optional<User> user = userDao.findById(loginUser.getUid());
-        if(!user.isPresent()) {
-            return new ResponseEntity<>("id", HttpStatus.BAD_REQUEST);
-        }
-        String encodePassword = passwordEncoder.encode(loginUser.getPassword());
-        if(passwordEncoder.matches(encodePassword, user.get().getPassword())) {
-            // 로그인 처리 jwt토큰 발급하기
-            // String token = jwtUtil.generateToken(user.get());
-            // return new ResponseEntity<>(token, HttpStatus.OK);
+    public Object Login(@RequestBody User loginUser) throws Exception {
+        authenticate(loginUser.getUid(), loginUser.getPassword());
+        final UserDetails userDetails = userDetailsService
+                .loadUserByUsername(loginUser.getUid());
+                
+        if(passwordEncoder.matches(loginUser.getPassword(), userDetails.getPassword())) {
+            final String token = jwtTokenUtil.generateToken(userDetails);
+            return new ResponseEntity<>(new JwtResponse(token), HttpStatus.OK);
         }
         return new ResponseEntity<>("password", HttpStatus.BAD_REQUEST);
     }
@@ -65,21 +78,26 @@ public class UserController {
         String encodePassword = passwordEncoder.encode(user.getPassword());
         user.setPassword(encodePassword);
         userDao.save(user);
-        // jwt 토큰 발급하기
-        return new ResponseEntity<>("ok", HttpStatus.OK);
+        final UserDetails userDetails = userDetailsService
+                .loadUserByUsername(user.getUid());
+        final String token = jwtTokenUtil.generateToken(userDetails);
+        return new ResponseEntity<>(new JwtResponse(token), HttpStatus.OK);
     }
 
     @GetMapping(value="/logout")
     public Object Logout() {
-
+        //TO DO
         return new ResponseEntity<>("ok", HttpStatus.OK);
     }
 
-    @GetMapping(value = "/test")
-    public Object test() {
-
-        return new ResponseEntity<>("test", HttpStatus.OK);
-    }
-    
+    private void authenticate(String username, String password) throws Exception {
+		try {
+			authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(username, password));
+		} catch (DisabledException e) {
+			throw new Exception("USER_DISABLED", e);
+		} catch (BadCredentialsException e) {
+			throw new Exception("INVALID_CREDENTIALS", e);
+		}
+	}
 
 }
